Extraction RdvBuilderConcern
